requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0
numpy>=1.24.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0

langchain==0.3.21
langchain-core==0.3.46
langchain-experimental==0.3.4
langchain-community==0.3.20
langchain-openai==0.3.9
langchain-anthropic==0.3.10
langchain-text-splitters==0.3.7
langchain-elasticsearch==0.3.2
langchain-cohere==0.4.3
langchain-milvus==0.1.8
langchain-google-genai==2.1.0
langchain-huggingface==0.1.2
langchain-azure-ai==0.1.2
langchain-teddynote==0.3.44
langchainhub==0.1.21
langgraph==0.3.18
langsmith==0.3.18
pinecone==5.4.2

#langchain-upstage==0.3.1
#langchain-upstage-community==0.3.1
#langchain-upstage-community-models==0.3.1

seaborn==0.13.2
matplotlib>=3.7.0
Pillow>=9.5.0

python-dotenv==1.0.1

tabulate==0.9.0

scipy>=1.10.0
scikit-learn>=1.3.0
openpyxl>=3.1.0